Moved all explanation fixtures to conftest
